Removes scenarioext_functions.cpp and merges its contents into scenarioext.cpp.
