--- src/extensions/scenario/scenarioext_functions.h
@@ -1,32 +1,0 @@
-/*******************************************************************************
-/*                 O P E N  S O U R C E  --  V I N I F E R A                  **
-/*******************************************************************************
- *
- *  @project       Vinifera
- *
- *  @file          SCENARIOEXT_FUNCTIONS.H
- *
- *  @author        CCHyper
- *
- *  @brief         Contains the supporting functions for the extended ScenarioClass.
- *
- *  @license       Vinifera is free software: you can redistribute it and/or
- *                 modify it under the terms of the GNU General Public License
- *                 as published by the Free Software Foundation, either version
- *                 3 of the License, or (at your option) any later version.
- *
- *                 Vinifera is distributed in the hope that it will be
- *                 useful, but WITHOUT ANY WARRANTY; without even the implied
- *                 warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
- *                 PURPOSE. See the GNU General Public License for more details.
- *
- *                 You should have received a copy of the GNU General Public
- *                 License along with this program.
- *                 If not, see <http://www.gnu.org/licenses/>.
- *
- ******************************************************************************/
-#pragma once
-
-
-void Vinifera_Assign_Houses();
-void Vinifera_Create_Units(bool official);
